services/inference-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
kafka-python==2.0.2
numpy==1.26.4
tritonclient[grpc]==2.41.0
